Iteration 6 Wave 1: EPVS (Epoch-Protected Version Scheme) core implementation.
Foundation for revivification and EPVS integration tests.

+++ rust/crates/faster-core/src/checkpoint/orchestrator.rs
@@ -646,7 +646,7 @@ mod tests {
         let meta = store.read_checkpoint_metadata(&result_token).unwrap();
 
         // Index recovery info
-        assert_eq!(meta.index_info.version, index.version());
+        assert_eq!(meta.index_info.version, index.version() as u64);
         assert_eq!(meta.index_info.table_size, index.num_buckets());
 
         // Log recovery info
